Add one file per command for methods
