+++ commands/methods_get.go
@@ -1,0 +1,117 @@
+package commands
+
+import (
+	"fmt"
+
+	"github.com/VictorAvelar/mollie-api-go/v2/mollie"
+	"github.com/VictorAvelar/mollie-cli/commands/displayers"
+	"github.com/avocatl/admiral/pkg/commander"
+	"github.com/avocatl/admiral/pkg/display"
+	"github.com/avocatl/admiral/pkg/prompter"
+	"github.com/spf13/cobra"
+)
+
+func getPaymentMethodCmd(p *commander.Command) {
+	gm := commander.Builder(
+		p,
+		commander.Config{
+			Namespace: "get",
+			ShortDesc: "Retrieve a single method by its ID.",
+			LongDesc: `Retrieve a single method by its ID. Note that if a method is not available on the website profile
+a status 404 Not found is returned. When the method is not enabled,a status 403 Forbidden
+is returned. You can enable payments methods via the Enable payment method endpoint in the
+Profiles API, or via your Mollie Dashboard.`,
+			Execute: getPaymentMethodAction,
+			Example: "mollie methods get --id=creditcard --locale=pt_PT",
+		},
+		methodsCols,
+	)
+
+	AddIDFlag(gm, true)
+	AddLocaleFlag(gm)
+	AddCurrencyCodeFlag(gm)
+}
+
+type getMethodPropmter struct {
+	Locale  string
+	Include string
+}
+
+func getPaymentMethodAction(cmd *cobra.Command, args []string) {
+	id, err := cmd.Flags().GetString(IDArg)
+	if err != nil {
+		logger.Fatal(err)
+	}
+
+	var opts mollie.MethodsOptions
+	{
+		if ParsePromptBool(cmd) {
+			v, err := prompter.Struct(&getMethodPropmter{})
+			if err != nil {
+				logger.Fatal(err)
+			}
+			val := v.(*getMethodPropmter)
+
+			opts.Locale = mollie.Locale(val.Locale)
+			opts.Include = val.Include
+		} else {
+			opts.Locale = mollie.Locale(ParseStringFromFlags(cmd, LocaleArg))
+			opts.Currency = ParseStringFromFlags(cmd, CurrencyArg)
+			opts.Include = ParseStringFromFlags(cmd, IncludeArg)
+
+		}
+	}
+
+	if verbose {
+		PrintNonEmptyFlags(cmd)
+	}
+
+	m, err := API.Methods.Get(id, &opts)
+	if err != nil {
+		logger.Fatal(err)
+	}
+
+	if json {
+		PrintJsonP(m)
+	}
+
+	err = printer.DisplayMany(getMethodsDisplayables(m))
+	if err != nil {
+		logger.Fatal(err)
+	}
+}
+
+func getMethodsDisplayables(m *mollie.PaymentMethodInfo) []display.Displayable {
+	method := &displayers.MollieMethod{
+		PaymentMethodInfo: m,
+	}
+
+	var dp []display.Displayable
+	{
+		dp = append(dp, method)
+
+		if len(m.Issuers) > 0 {
+			if verbose {
+				dp = append(dp, displayers.NewSimpleTextDisplayer("=", fmt.Sprintf("Embeded issuers for method %s", m.Description)))
+			}
+			dp = append(dp, &displayers.MollieListPaymentMethodsIssuers{
+				Issuers: m.Issuers,
+			})
+		}
+	}
+
+	if verbose {
+		var text string
+		{
+			for c, d := range method.ColMap() {
+				text += fmt.Sprintf("%s:\t%v\n", c, d)
+			}
+		}
+
+		colmap := displayers.NewSimpleTextDisplayer("=", text)
+
+		dp = append([]display.Displayable{colmap}, dp...)
+	}
+
+	return dp
+}
